Update `_ensure_gitignore_entries()`.
There are two major implementation changes undertaken by this commit:

Instead of opening the file three times ('a', 'r', 'w'), we now open the file
just once (with mode 'a+') for appending and updating, which is sufficient to
read the `.gitignore` for matching entries and then append any missing entries
to the end of the file. The file is created if missing, not truncated if
present, is seekable, readable, and "append"able. Note that, per the Python docs
on open[^1], all writes to a file object opened with 'a' "append to the end of
the file regardless of the current seek position" on some Unix systems; however,
we position the seek position at the end of the file regardless (after reading
the file contents), so the two cases are unified into desired behavior.

Appending only the missing lines to the `.gitignore` file is not only slightly
faster (because we don't need to open/close three file objects and don't need to
truncate and rewrite an arbitrarily large pre-existing `.gitignore`), but also
slightly more correct, because `.gitignore` respects trailing spaces escaped by
`\` (e.g. `/foo\ `)[^2], but by writing trimmed entries back to the file, we
were previously not respecting pre-existing `.gitignore` entries of that form.
Truthfully, I can't imagine that very many repositories include files that end
with trailing spaces (quite odd) AND care enough to explicitly ignore those
files with this unusual `.gitignore` format, but we now respect those entries.

As usual, this commit also includes additional clarifying comments encountered
as I was parsing the implementation.

This does _not_ change the explainer comments alongside the patterns, because
that technically changes the interface of this tool, but I would recommend
updating those explainer comments to be more explanatory (in line with the local
comments in this repository's `.gitignore`.)

For consistency with `_ensure_external_workspaces_link_exists()`, this also
renames `_ensure_gitignore_entries()` to `_ensure_gitignore_entries_exist()`.

[^1]: https://docs.python.org/3/library/functions.html#open
[^2]: https://git-scm.com/docs/gitignore#_pattern_format